+++ openstack/blockstorage/extensions/volumeactions/begin_detaching.go
@@ -2,10 +2,8 @@ package volumeactions
 
 import "github.com/opentelekomcloud/gophertelekomcloud"
 
-func BeginDetaching(client *golangsdk.ServiceClient, id string) (r BeginDetachingResult) {
+func BeginDetaching(client *golangsdk.ServiceClient, id string) (err error) {
 	b := map[string]interface{}{"os-begin_detaching": make(map[string]interface{})}
-	raw, err := client.Post(client.ServiceURL("volumes", id, "action"), b, nil, &golangsdk.RequestOpts{
-		OkCodes: []int{202},
-	})
+	_, err = client.Post(client.ServiceURL("volumes", id, "action"), b, nil, nil)
 	return
 }
