+++ openstack/blockstorage/extensions/volumeactions/force_delete.go
@@ -4,7 +4,7 @@ import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
 )
 
-func ForceDelete(client *golangsdk.ServiceClient, id string) (r ForceDeleteResult) {
-	raw, err := client.Post(client.ServiceURL("volumes", id, "action"), map[string]interface{}{"os-force_delete": ""}, nil, nil)
+func ForceDelete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Post(client.ServiceURL("volumes", id, "action"), map[string]interface{}{"os-force_delete": ""}, nil, nil)
 	return
 }
